Move Directionality and NumberOfDirections to Native
